docs: :pencil: 0.9.0
