+++ example.ts
@@ -1,0 +1,15 @@
+import {
+  UpdateNotifier,
+  Github,
+} from "./mod.ts";
+
+const notifier = new UpdateNotifier({
+  name: "denon",
+  owner: "denosaurs",
+  registry: Github,
+  currentVersion: "0.1.2",
+  updateCheckInterval: 0,
+});
+
+await notifier.checkForUpdates();
+notifier.notify();
